Clock in, roster and book absence on the engine the box now runs
packages/workforce was converted by no step in the plan, and nothing could see it: the guard against
leftover PostgreSQL SQL scanned apps/server only. Seventeen of its twenty-two test files were
failing, and with them the seventeen workforce routes and the twelve schedule and me routes — `near
"at": syntax error` for shifts and swaps, `unrecognized token: ":"` for absences.

The package is 16 files of 22 now, 275 tests passing where 153 did: 122 red to green, NONE green to
red, diffed by sorted test name over the same 299 tests. The three consumer suites in apps/server run
the real HTTP routes against a real venue file with no mocks, and they went from 31 failing to 117
passing.

The shift's local wall date — a timestamp shifted by its own stored offset, then taken as a day — had
been written out three times. It has one home now, and the substitution was measured in both
directions: a positive offset moves the day forward, a negative one back, and a zero offset leaves
the UTC day, which is the control that says the shift is the offset's doing and not the function's.

FOUR LOSSES, each named where it happens. A read-back no longer normalises the stored spelling of a
timestamp, because the formatting call that did that is gone; callers accept anything Date.parse
accepts, `+02:00` included. A date built from an unparseable instant is now NULL rather than an
error, so such a row drops out of a window silently instead of failing the query. A pragma's table
name is written into the statement rather than bound, because this engine refuses to bind it — the
names are the file's own constants. And the interval check on a shift is no longer an interval check
when its two endpoints are spelled with different offsets: measured with four inserts and a control
in each direction, it ACCEPTS an interval that ends an hour before it starts, and REFUSES a valid
one-hour shift. The first never reaches the database, because the application compares instants
before writing; the second does, and answers a raw constraint failure instead of the domain refusal.
The fix is one line but it changes what a public read returns, so it is a decision and it is in the
plan rather than taken here.

Two assertions came out STRONGER. Where the old engine's refusal carried a class and no name, this
one names the constraint, so two enum cases now pin the constraint they mean. And an assertion
matching a partial unique index BY NAME could never have passed here — the engine names the index's
COLUMNS — so it matches those instead.

The residue guard gains this package as a root AND three patterns, which is the part worth keeping:
run the existing five patterns over the pre-change chain.ts and they report ZERO offenders, while the
statement in it was refused at prepare. Widening the roots without widening the patterns would have
missed the exact residue the widening was for. Three controls say the widening is real — planted
residue with the root removed passes, with the root restored fails naming the file and line, and with
the two new patterns removed passes again. The controller re-ran the middle one independently.

Left failing, unchanged and not this change's: two cases in the file that exists to record that
PGlite cannot test lock contention, and five real-PostgreSQL suites that fail at their hook for want
of the harness this branch removed — 145 files across the workspace are in that position. Two of the
five matter more than the rest and are named in the plan: one proves the append-only floor through a
role's privileges, and one is the §5 restore-continuation guard.

Also recorded in the plan, from the same work: `pgErrorCode` answers one code string for every
failure on this engine, so every surviving SQLSTATE assertion in the tree is either red or vacuous;
and a stubbed rejection has to carry `errcode`, not `code` — two cases here forged the latter, so the
collision they staged was not one and the retry under test never ran.

+++ packages/workforce/src/absences.test.ts
@@ -114,7 +114,7 @@ describe("createAbsence", () => {
       }),
     );
     const rows = await suite.db.execute<{ n: number }>(
-      sql`select count(*)::int as n from absences where id = ${id}`,
+      sql`select count(*) as n from absences where id = ${id}`,
     );
     expect(rows.rows[0]!.n).toBe(1);
   });
@@ -156,7 +156,7 @@ describe("createAbsence", () => {
       }),
     );
     const rows = await suite.db.execute<{ n: number }>(
-      sql`select count(*)::int as n from absences where id = ${id}`,
+      sql`select count(*) as n from absences where id = ${id}`,
     );
     expect(rows.rows[0]!.n).toBe(1);
   });
@@ -181,7 +181,7 @@ describe("createAbsence", () => {
       }),
     );
     const rows = await suite.db.execute<{ n: number }>(
-      sql`select count(*)::int as n from absences where id = ${id}`,
+      sql`select count(*) as n from absences where id = ${id}`,
     );
     expect(rows.rows[0]!.n).toBe(1);
   });
